% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{connectDb}
\alias{connectDb}
\title{Create a connection to a PostgreSQL database}
\usage{
connectDb(
  host = "localhost",
  port = "5432",
  dbname = "platform",
  user = "postgres",
  password = "password"
)
}
\arguments{
\item{host}{A character string indicating the host name or IP address of the
server on which the database is running. Default is “localhost".}

\item{port}{A character string indicating the port number on which to connect
to the database. Default is “5432".}

\item{dbname}{A character string indicating the name of the PostgreSQL
database to connect to.}

\item{user}{A character string indicating the user name to use when
connecting to the database.}

\item{password}{A character string indicating the password to use when
connecting to the database.}
}
\value{
A list object containing the following elements:
\itemize{
\item \code{query}: A function used to execute SQL queries.
\item \code{insert}: A function used to insert data into tables.
\item \code{disconnect}: A function used to disconnect from the database.
}
}
\description{
This function creates a connection to a PostgreSQL database and returns it as
a list object with three functions - \code{query}, \code{insert}, and
\code{disconnect}. These functions can be called to execute SQL queries,
insert data into tables and disconnect from the database respectively.
}
