% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{save}
\alias{save}
\title{Save query results as RDS file}
\usage{
save(data, name = NULL, dir = "~/shared/rds", force = FALSE)
}
\arguments{
\item{data}{A list object returned by the \code{\link{query}} function.}

\item{name}{An optional character string used as the name of the output file.
If not provided, the file name will be generated using the CRC32 hash
algorithm on the query statement.}

\item{dir}{An optional character string indicating the directory in which to
save the file. Default is “rds”.}

\item{force}{A logical value indicating whether to overwrite an existing file.
Default is FALSE.}
}
\value{
This function has no explicit return value.
It saves the RDS file to the specified directory.
}
\description{
This function saves query results as an RDS file.
The name of the RDS file is generated using the CRC32 hash algorithm on the
query statement, unless a name is provided.
The file is saved in the specified directory unless the directory does not
exist or the file already exists, in which case the function will print out
an appropriate message.
}
\examples{
\dontrun{
# Save query results to an RDS file
vaxr::save(my_query, "my_query_results")
}

}
