% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{upload}
\alias{upload}
\title{Insert data into a PostgreSQL database table}
\usage{
upload(data, table, db = NULL, append = TRUE, row.names = FALSE)
}
\arguments{
\item{data}{A data frame containing the data to be inserted into the database.}

\item{table}{A character string indicating the name of the PostgreSQL database
table to insert data into.}

\item{db}{An optional PostgreSQL database connection object. If not provided,
the function will connect to the default database.}

\item{append}{A logical value indicating whether to append the data to an
existing table or create a new table. Default is TRUE.}

\item{row.names}{A logical value indicating whether to preserve row names.
Default is FALSE.}
}
\value{
This function has no explicit return value.
}
\description{
This function inserts data into a PostgreSQL database table.
The function takes a data frame, table name, and optional database connection
object as input. If no database connection object is provided, the function
will create a new connection object. The function also accepts optional
arguments for appending data to the existing table and preserving row names.
}
\examples{
\dontrun{
# Insert data into a table
vaxr::upload(my_data, "my_table")
}

}
